PR comments
- Use _EPHEMERAL attribute
- Move tests to separate file
- Docstring updates
